fix(security): close remaining response sinks
