+++ backend/api/routers/engines.py
@@ -15,9 +15,9 @@ Environment variables (`OMNIVOICE_TTS_BACKEND`, `OMNIVOICE_ASR_BACKEND`,
 `OMNIVOICE_LLM_BACKEND`) still win over the UI choice so power-users can pin
 a backend without Settings silently undoing it.
 """
+import logging
 import os
 import threading
-import logging
 from time import perf_counter
 
 from fastapi import APIRouter, Depends, HTTPException
@@ -335,7 +335,6 @@ def engine_health(engine_id: str):
         )
 
     t0 = perf_counter()
-    failure_class = "HealthCheckFailed"
     if hasattr(cls, "health_check"):
         # SubprocessBackend path — spawn sidecar (if not running) and ping.
         # ``health_check`` already swallows its own exceptions per Plan
@@ -345,7 +344,6 @@ def engine_health(engine_id: str):
             instance = _get_engine_instance(cls)
             ok, msg = instance.health_check()
         except Exception as exc:
-            failure_class = type(exc).__name__
             ok, msg = False, f"{type(exc).__name__}: {exc}"
     else:
         # In-process backend — `is_available()` is the classmethod-level
@@ -354,20 +352,15 @@ def engine_health(engine_id: str):
         try:
             ok, msg = cls.is_available()
         except Exception as exc:
-            failure_class = type(exc).__name__
             ok, msg = False, f"{type(exc).__name__}: {exc}"
 
     # Engine-owned output can contain much more than shaped HF tokens: local
     # paths, arbitrary credentials, source lines, or a nested traceback.
-    from core.response_safety import public_engine_health
+    from core.public_errors import public_engine_health
 
     latency_ms = (perf_counter() - t0) * 1000.0
     if not ok:
-        logger.warning(
-            "Engine %s health check failed (class=%s; details withheld)",
-            engine_id,
-            failure_class,
-        )
+        logger.warning("Engine health check failed; details withheld")
     return {
         "id": engine_id,
         "ok": bool(ok),
