+++ backend/core/public_errors.py
@@ -1,6 +1,9 @@
 """Data-independent error metadata safe for API and streaming responses."""
 from __future__ import annotations
 
+import logging
+from typing import Any
+
 _PROVIDER_DETAILS = {
     "auth": "Authentication failed. Check the provider API key.",
     "not_found": "Provider or model not found. Check the model and Base URL.",
@@ -51,3 +54,51 @@ def stream_failure(code: str) -> dict[str, object]:
         },
     }
     return dict(failures.get(code, failures["generation_failed"]))
+
+
+def public_failure(
+    logger: logging.Logger,
+    log_message: str,
+    error: BaseException | object,
+    *,
+    response: str,
+    traceback: bool = False,
+) -> str:
+    """Log fixed failure metadata and return a fixed public failure message.
+
+    ``response`` must be authored by VoiceStudio, never derived from ``error``.
+    The helper intentionally does not attempt to redact exception text: a
+    deny-list cannot cover arbitrary secrets, paths, source lines or nested
+    tracebacks.
+    """
+    del traceback
+    error_class = type(error).__name__ if isinstance(error, BaseException) else "Failure"
+    logger.error("%s (class=%s; details withheld)", log_message, error_class)
+    return response
+
+
+def public_engine_health(ok: bool, diagnostic: Any) -> str:
+    """Map an engine-owned health diagnostic to a stable response message."""
+    del diagnostic
+    return "Healthy" if ok else "Engine unavailable; check the backend log for details."
+
+
+def public_exception_response(error: BaseException, *, fallback: str) -> dict[str, str]:
+    """Return fixed remediation selected by a stable failure taxonomy.
+
+    Classification may inspect the private diagnostic locally, but response
+    values come exclusively from VoiceStudio-owned constants. No substring of
+    ``error`` is copied into the payload.
+    """
+    from core.failure import classify, public_hint_for_topic
+
+    try:
+        topic = classify(str(error))
+        hint = public_hint_for_topic(topic)
+    except Exception:
+        topic = ""
+        hint = ""
+    payload = {"detail": f"{fallback} {hint}".strip()}
+    if topic and hint:
+        payload.update({"docs_topic": topic, "hint": hint})
+    return payload
