+++ backend/main.py
@@ -966,7 +966,7 @@ async def global_exception_handler(request: Request, exc: Exception):
     # Appending the shared hints HERE covers every route that can leak a
     # model-load/download error (generate, dub, archetypes, …), not just TTS
     # generate. append_hint is a no-op for every other error and never raises.
-    from core.response_safety import public_exception_response
+    from core.public_errors import public_exception_response
 
     content = public_exception_response(
         exc,
